+++ common/src/main/java/com/st0x0ef/stellaris/common/armors/AbstractSpaceArmor.java
@@ -26,7 +26,7 @@ public abstract class AbstractSpaceArmor extends CustomArmorItem {
         }
 
         public void onArmorTick(ItemStack stack, Level level, Player player) {
-            if (!PlanetUtil.hasOxygen(level.dimension().location()) && !player.isCreative()) {
+            if (!PlanetUtil.hasOxygen(level) && !player.isCreative()) {
                 OxygenUtils.removeOxygen(stack, FluidStack.bucketAmount()/1000);
             }
         }